pyproject.toml
[tool.poetry]
name = "twitfaver"
version = "0.1.0"
description = "Twitter favoriter"
authors = ["Batista Harahap <batista@bango29.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
APScheduler = "^3.8.1"
python-twitter = "^3.5"
nltk = "^3.7"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[tool.black]
line-length = 120
target-version = ["py310"]
skip-string-normalization = true
skip-magic-trailing-comma = true
exclude = '''
    /(
        \.git
      | \.hg
      | \.mypy_cache
      | \.tox
      | \.venv
      | _build
      | buck-out
      | build
      | dist
      # The following are specific to Black, you probably don't want those.
      | blib2to3
      | tests/data
    )/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
